services/browser-use/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
browser-use>=0.2.0
langchain-openai>=0.3.0
pydantic>=2.10.0
python-dotenv>=1.0.0
